filteroptions for linkfield
