+++ src/features/payload-cms/payload-cms/shared-blocks/form-block.ts
@@ -1,33 +1,5 @@
-import type { Locale } from '@/types/types';
-import type { Block, FilterOptionsProps, Where } from 'payload';
-
-const findQuery: ({ req, relationTo }: FilterOptionsProps<unknown>) => Promise<Where> = async ({
-  req,
-  relationTo,
-}) => {
-  const { payload, locale } = req;
-
-  const allItems = await payload.find({
-    collection: relationTo,
-    draft: false,
-    locale: locale as Locale,
-    limit: 10,
-    where: {
-      _localized_status: {
-        equals: {
-          published: true,
-        },
-      },
-    },
-    req,
-  });
-  const allIds = allItems.docs.map((item) => item.id);
-  return {
-    id: {
-      in: allIds,
-    },
-  };
-};
+import { filterOptionsOnlyPublished } from '@/features/payload-cms/payload-cms/utils/filter-options-only-published';
+import type { Block } from 'payload';
 
 export const formBlock: Block = {
   slug: 'formBlock',
@@ -43,7 +15,7 @@ export const formBlock: Block = {
       relationTo: 'forms',
       required: true,
       hasMany: false,
-      filterOptions: findQuery,
+      filterOptions: filterOptionsOnlyPublished,
       validate: () => true,
     },
   ],
